// src/types/Complaint.ts
export interface ComplaintData {
    company: string;
    company_public_response: string;
    company_response: string;
    complaint_id: number;
    complaint_what_happened: string;
    consumer_consent_provided: string;
    consumer_disputed: string;
    date_received: string;
    date_sent_to_company: string;
    has_narrative: boolean;
    issue: string;
    product: string;
    state: string;
    sub_issue: string;
    sub_product: string;
    submitted_via: string;
    tags: string;
    timely: string;
    zip_code: string;
  }